+++ memberaudit_securegroups/apps.py
@@ -1,7 +1,16 @@
+"""
+App's base definitions
+"""
+
+# Django
 from django.apps import AppConfig
 
 
 class MemberAuditSecureGroupsConfig(AppConfig):
+    """
+    MemberAuditSecureGroupsConfig
+    """
+
     name = "memberaudit_securegroups"
     label = "memberaudit_securegroups"
     verbose_name = "Member Audit Secure Groups Integration"