+++ memberaudit_securegroups/auth_hooks.py
@@ -1,6 +1,12 @@
+"""
+Hook into Alliance Auth
+"""
+
+# Alliance Auth
 from allianceauth import hooks
 
-from .models import (
+# Memberaudit Securegroups
+from memberaudit_securegroups.models import (
     ActivityFilter,
     AgeFilter,
     AssetFilter,